feat : read and insert benchmark
test_insert : take the file in parameter, given by the main which get it
through his argument (argv). Only use it with the filesystem that
provides the right API to insert string into file

test_read : take 2 files in paramater, the first file should be into a
partion that have been formated with working fs, and the second file
located into our fs. It will read at the same random place (the 2 files
should contain the exact same thing), and if our fs works, we should
read the same thing.