Revert "feat:user ouichfs  write" (#9)
* Revert "feat:user ouichfs  write (#8)"

This reverts commit 6dff8490cf92aba108ecdd703254317a68b443ea.

* chores: moved benchmark.c

* feat: benchmark wrtite user

* feat: user write benchmark

* feat: user write benchmark

To be use only after creating the kernel API for lite insertion file
system because the current kernel API cannot perform a string insertion,
we need to do it manually.

* feat : read and insert benchmark

test_insert : take the file in parameter, given by the main which get it
through his argument (argv). Only use it with the filesystem that
provides the right API to insert string into file

test_read : take 2 files in paramater, the first file should be into a
partion that have been formated with working fs, and the second file
located into our fs. It will read at the same random place (the 2 files
should contain the exact same thing), and if our fs works, we should
read the same thing.

* feat : user read benchmark

* fix: user read benchmark

* fix: get right bloc size

To get the right bloc size that matches the ouichfs specification

* fix: use of clock_gettime